splitting csv parsing stuff into somewhat pluggable modules
