+++ go/contacts/templates/contacts/import_failed_mail.txt
@@ -12,6 +12,7 @@ Vumi Go.
 
 Account Key: {{account_key}}
 Group Key: {{group_key}}
-CSV Path: {{csv_path}}
+File Type: {{file_type}}
+File Path: {{file_path}}
 Field Names: {{field_names}}
 Has Header: {{has_header}}
